Cleanup Remove Audit, add AuditConfig. Put EstimFn with AlphaMart. Improve MultiContestTestData, remove MakeRandomTestData.
