+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/sampling/StylishWorkflow.kt
@@ -9,19 +9,6 @@ import java.util.concurrent.TimeUnit
 
 private val showQuantiles = false
 
-data class AuditConfig(val auditType: AuditType,
-                       val riskLimit: Double,
-                       val seed: Long,
-                       val ntrials: Int = 100,
-                       val quantile: Double = .80,
-                       val p1: Double = 1.0e-2,
-                       val p2: Double = 1.0e-4,
-                       val p3: Double = 1.0e-2,
-                       val p4: Double = 1.0e-4,
-                       val d1: Int = 100,  // for trunc_shrinkage
-                       val d2: Int = 100,
-    )
-
 // STYLISH 2.1
 //Card-level Comparison Audits and Card-Style Data
